prefect/functions/parse-rss/requirements.txt
functions-framework==3.*
google-cloud-storage
google-cloud-secret-manager
duckdb==1.1.0
pandas
beautifulsoup4